+++ packages/tdesign-miniprogram/example/pages/home/data/base.ts
@@ -43,10 +43,6 @@ const base = {
     //   name: 'Typography',
     //   label: '字体',
     // },
-    {
-      name: 'ConfigProvider',
-      label: '全局特性配置',
-    },
   ],
 };